feat(cloud): 补全 16 个 CS→CP handler 的 HTTP 分支 + 清理 DispatchError
- 17 个 CS→CP handler 全部实现 message_by_http 分支（除已在 31b7d10 实现的 cancel_reservation，本次填实 16 个）

- 统一响应解析模式：拆分 CALLRESULT / CALLERROR / 未知类型三种结果

- 调用方错误信息携带 unique_id 便于排查

- ocpp-1-6 protocol: CALL/CALLRESULT/CALLERROR 改 i64，与 serde_json::Value::as_i64() 对齐

- ocpp16/kafka/dispatcher.rs: 删除冗余 DispatchError enum，统一用 AppError

- ocpp16/envelope.rs: CloudMessage 新增 http_url 字段（HTTP 出站需要）

- infra.rs 注册 http_client 模块；infra/http_client.rs 小幅调整

- Cargo.toml default feature 改回 message_by_http（与 README/deploy 默认一致）

+++ crates/ocpp-1-6/src/protocol.rs
@@ -9,11 +9,11 @@ pub const OCPP_VERSION: &str = "1.6";
 
 // 消息类型 ID（用于消息首位标识）
 /// CALL 消息类型 ID（请求）
-pub const CALL: i32 = 2;
+pub const CALL: i64 = 2;
 /// CALLRESULT 消息类型 ID（成功响应）
-pub const CALLRESULT: i32 = 3;
+pub const CALLRESULT: i64 = 3;
 /// CALLERROR 消息类型 ID（错误响应）
-pub const CALLERROR: i32 = 4;
+pub const CALLERROR: i64 = 4;
 
 // 操作名称：常用的 OCPP 操作字符串常量，便于避免魔法字符串
 pub const ACTION_AUTHORIZE: &str = "Authorize";
